requirements.txt
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
django-filter==22.1
requests==2.28.1